Use lowest finger instead of earliest finger to determine button.
Previously the button emulation code used the finger that was placed on the
pad first. For example, if you moved the cursor with your index finger
and then tried to press a button with thumb, it would determine the button
from the location of the index finger.

This patch changes it to use the lowest finger currently on the touchpad.
This works as long as the button press area is at the bottom of the touchpad,
which seems like a fair assumption. Also "invalid" touches are taken into
account, as otherwise the "10% button area" filters out just the finger
we are interested in.